Separated KeySignSection and GetKeySection into two independent files
